requirements.txt
flask
pymysql
pyserial